+++ ethereum/node.go
@@ -27,31 +27,3 @@ func NewEthConfig(ctx *cli.Context, stack *node.Node) *eth.Config {
 
 	return eth_config
 }
-
-// MakeSystemNode sets up a local node and configures the services to launch
-// func MakeSystemNode(name, version string, ctx *cli.Context) *node.Node {
-
-// 	// Sets the target gas limit
-// 	ethUtils.SetupNetwork(ctx)
-
-// 	// Setup the node, a container for services
-// 	// TODO: dont think we need a node.Node at all
-// 	nodeConf := NewNodeConfig(ctx)
-// 	stack, err := node.New(nodeConf)
-// 	if err != nil {
-// 		ethUtils.Fatalf("Failed to create the protocol stack: %v", err)
-// 	}
-
-// 	// Configure the eth
-// 	ethConf := NewEthConfig(ctx, stack)
-
-// 	//Remote tendermint RPC address
-// 	tendermintURI := ctx.GlobalString(emtUtils.BroadcastTxAddrFlag.Name)
-
-// 	if err := stack.Register(func(ctx *node.ServiceContext) (node.Service, error) {
-// 		return NewBackend(ctx, ethConf, rpcclient.NewClientURI(tendermintURI))
-// 	}); err != nil {
-// 		ethUtils.Fatalf("Failed to register the ABCI application service: %v", err)
-// 	}
-// 	return stack
-// }
